The log file location was changed from /tmp to /var/log, according to a logrotate security requirements
When a log file locates in a `/tmp`, the `logrotate` service has failed with error message `error: skipping "/tmp/nats-server.log" because parent directory has insecure permissions (It's world writable or writable by group which is not "root")`. Replacing file to the common log directory `/var/log` solve this issue.