Improve JS concepts
Looking at a HN thread you can tell a lot of people put systems in two categories: broker based (eg. Rabbit, et al) that provide proper queueing (but no replay) and 'streaming' (Kafka) that provides replay (but doesn't do proper queueing).
So highlighting the fact that NATS is unique in that it covers both because we have DQ of course but more because we have working-queue streams (and the proper many different acks and ack per message). Same goes for sourcing being built-in to streams.